Clean up, restructuring and renaming
