+++ dynomitemanager/src/main/java/com/netflix/dynomitemanager/sidecore/backup/S3Restore.java
@@ -15,9 +15,6 @@ package com.netflix.dynomitemanager.sidecore.backup;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
-import java.text.SimpleDateFormat;
-import java.util.Calendar;
-import java.util.Date;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -33,7 +30,6 @@ import com.netflix.dynomitemanager.sidecore.ICredential;
 import com.netflix.dynomitemanager.identity.InstanceIdentity;
 import com.amazonaws.AmazonClientException;
 import com.amazonaws.AmazonServiceException;
-import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.model.GetObjectRequest;
 import com.amazonaws.services.s3.model.S3Object;
